requirements.txt
Flask==3.1.2
flask-cors==6.0.2
gunicorn==25.0.1

psycopg2-binary==2.9.11
python-dotenv==1.2.1

requests==2.32.5
httpx==0.28.1

beautifulsoup4==4.14.3
lxml==6.0.2
trafilatura==2.0.0

google-genai==1.61.0
tenacity==9.1.2

ddgs
lxml_html_clean
python-slugify